[move] ./moves/ folder renaming
