+++ c++/moves/double_insert.hpp
@@ -19,14 +19,13 @@
  *
  ******************************************************************************/
 #pragma once
-#include "./qmc_data.hpp"
-#include <algorithm>
+#include "../qmc_data.hpp"
 #include <triqs/mc_tools.hpp>
 
 namespace cthyb {
 
-// Removal of 2 C and 2 C^dagger operator
-class move_remove_c_c_cdag_cdag {
+// Insertion of 2 C and 2 C^dagger operators
+class move_insert_c_c_cdag_cdag {
 
  qmc_data& data;
  configuration& config;
@@ -38,11 +37,12 @@ class move_remove_c_c_cdag_cdag {
  double dtau1, dtau2;
  h_scalar_t new_atomic_weight, new_atomic_reweighting;
  time_pt tau1, tau2, tau3, tau4;
+ op_desc op1, op2, op3, op4;
 
  histogram* add_histo(std::string const& name, histo_map_t* histos);
 
  public:
- move_remove_c_c_cdag_cdag(int block_index1, int block_index2, int block_size1, int block_size2, std::string const& block_name1,
+ move_insert_c_c_cdag_cdag(int block_index1, int block_index2, int block_size1, int block_size2, std::string const& block_name1,
                            std::string const& block_name2, qmc_data& data, mc_tools::random_generator& rng, histo_map_t* histos)
     : data(data),
       config(data.config),
@@ -51,10 +51,10 @@ class move_remove_c_c_cdag_cdag {
       block_size1(block_size1),
       block_index2(block_index2),
       block_size2(block_size2),
-      histo_proposed1(add_histo("double_remove_length_proposed_" + block_name1, histos)),
-      histo_proposed2(add_histo("double_remove_length_proposed_" + block_name2, histos)),
-      histo_accepted1(add_histo("double_remove_length_accepted_" + block_name1, histos)),
-      histo_accepted2(add_histo("double_remove_length_accepted_" + block_name2, histos)) {}
+      histo_proposed1(add_histo("double_insert_length_proposed_" + block_name1, histos)),
+      histo_proposed2(add_histo("double_insert_length_proposed_" + block_name2, histos)),
+      histo_accepted1(add_histo("double_insert_length_accepted_" + block_name1, histos)),
+      histo_accepted2(add_histo("double_insert_length_accepted_" + block_name2, histos)) {}
 
  mc_weight_t attempt();
  mc_weight_t accept();
